[DDP][PT2D][2D] Enable DDP + TP and add test for compiled DDP + TP
Pull Request resolved: https://github.com/pytorch/pytorch/pull/120479

This PR enables DDP + TP using a TP internal API. This should not be the final implementation a more sound implementation is to inline the TP internal API in DDP. In other words, DDP needs to be aware of DTensor so that we can support 2D state_dict.

**TODOs**

- [x] Implement DDP allreduce fusion algorithm for Inductor post_grad pass.
- [x] Add unit tests to ensure the fusion doesn't DDP + TP.
- [ ] Group different PG and data type of all_reduces.
- [ ] Mixed precision supports and tests
- [ ] Implement the fusions with Inductor IR.
- [ ] Add auto bucketing based on Inductor profiling.


This PR adds a compiled DDP + TP test to ensure the new compiled DDP fusion doesn't break TP all_reduce.
ghstack-source-id: 217275459
@exported-using-ghexport

Differential Revision: [D54105050](https://our.internmc.facebook.com/intern/diff/D54105050/)